+++ src/lab/mcp/example/genai_client.py
