+++ src/lab/mcp/example/manual_client.py
