+++ src/lab/mcp/example/openai_client.py
